Use gradle to automate some of the build tasks
