--- src/main/java/command/ExitCommand.java
@@ -1,12 +1,0 @@
-package command;
-
-import storage.StorageOperationException;
-import task.TaskList;
-import ui.Ui;
-
-public class ExitCommand extends Command{
-    @Override
-    public void execute(TaskList list, Ui ui) {
-        ui.exit();
-    }
-}
