+++ src/main/java/mylo/data/NoSuchCommandException.java
@@ -1,0 +1,13 @@
+package mylo.data;
+
+public class NoSuchCommandException extends Exception {
+
+    public NoSuchCommandException(String input) {
+        super(String.format("%s is not a valid command. Please try again with: \ntodo \nevent \ndeadline \nlist \nbye", input));
+    }
+
+    @Override
+    public String toString() {
+        return super.getMessage();
+    }
+}
