+++ src/main/java/mylo/storage/FileCorruptedException.java
@@ -1,0 +1,8 @@
+package mylo.storage;
+
+public class FileCorruptedException extends Exception {
+
+    public FileCorruptedException(String message) {
+        super(message);
+    }
+}
