+++ src/main/java/mylo/storage/StorageOperationException.java
@@ -1,0 +1,10 @@
+package mylo.storage;
+
+/**
+ * Signals that there were errors converting and/or storing data to file.
+ */
+public class StorageOperationException extends Exception {
+    public StorageOperationException(String message) {
+        super(message);
+    }
+}
