+++ src/main/java/mylo/task/TaskType.java
@@ -1,0 +1,5 @@
+package mylo.task;
+
+public enum TaskType{
+    TODO, EVENT, DEADLINE
+}
