+++ src/main/java/mylo/utils/formatters/Formatter.java
@@ -1,0 +1,14 @@
+package mylo.utils.formatters;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Formatter {
+    public static String dateTimeDisplay(LocalDateTime dateTime) {
+        return dateTime.format(DateTimeFormatter.ofPattern("EEE dd MMM yyyy HH:mm"));
+    }
+
+    public static String dateTimeStorage(LocalDateTime dateTime) {
+        return dateTime.format(DateTimeFormatter.ofPattern("dd-LL-yyyy HHmm"));
+    }
+}
