--- src/main/java/storage/FileCorruptedException.java
@@ -1,8 +1,0 @@
-package storage;
-
-public class FileCorruptedException extends Exception {
-
-    public FileCorruptedException(String message) {
-        super(message);
-    }
-}
