--- src/main/java/storage/StorageOperationException.java
@@ -1,10 +1,0 @@
-package storage;
-
-/**
- * Signals that there were errors converting and/or storing data to file.
- */
-public class StorageOperationException extends Exception {
-    public StorageOperationException(String message) {
-        super(message);
-    }
-}
