--- src/main/java/task/TaskType.java
@@ -1,5 +1,0 @@
-package task;
-
-public enum TaskType{
-    TODO, EVENT, DEADLINE
-}
