--- src/main/java/utils/exceptions/IllegalValueException.java
@@ -1,13 +1,0 @@
-package utils.exceptions;
-
-/**
- * Signals that some given data does not fulfill some constraints.
- */
-public class IllegalValueException extends Exception {
-    /**
-     * @param message should contain relevant information on the failed constraint(s)
-     */
-    public IllegalValueException(String message) {
-        super(message);
-    }
-}
